skills/video-master/requirements.txt
Pillow>=10.4,<12
edge-tts>=7,<8
imageio[ffmpeg]>=2.34,<3
numpy>=1.26,<3
openpyxl>=3.1,<4
pydantic>=2.7,<3
pysubs2>=1.8,<2
